reworked class system + new neuron_object class and classification_library class to begin with neural networks
