+++ regression_library.py
@@ -1,7 +1,6 @@
 # This class was created by Bendik Arbogast at the 17.10.2020 and is available free of charge to the general public.
 # All rights reserved. If you have any questions or ideas to improve the contents of this file
 # please consider writing an email to arbobendik@gmail.com or contact me on GitHub.
-
 from typing import Callable
